+++ src/sub/read_assembly_file.h
@@ -33,7 +33,7 @@ int isValidLabelName(char *str);
 
 
 // Read a line of assembly code
-assemblyLine *readAssemblyLine(size_t number, char *line, size_t lineID);
+assemblyLine *readAssemblyLine(int number, char *line, int lineID);
 
 
 // Read an "assembly language" file and return a list of all the lines
